Refactor timegraph and ganttchart to extend basegantt
